refactor: extract selection detail facts helpers
